+++ sentry-android-okhttp/src/test/java/io/sentry/android/okhttp/SentryOkHttpInterceptorTest.kt
@@ -39,6 +39,7 @@ class SentryOkHttpInterceptorTest {
         var interceptor = SentryOkHttpInterceptor(hub)
         val server = MockWebServer()
         lateinit var sentryTracer: SentryTracer
+        lateinit var options: SentryOptions
 
         @SuppressWarnings("LongParameterList")
         fun getSut(
@@ -47,19 +48,19 @@ class SentryOkHttpInterceptorTest {
             responseBody: String = "success",
             socketPolicy: SocketPolicy = SocketPolicy.KEEP_OPEN,
             beforeSpan: SentryOkHttpInterceptor.BeforeSpanCallback? = null,
-            includeMockServerInTracingOrigins: Boolean = true
+            includeMockServerInTracePropagationTargets: Boolean = true,
+            keepDefaultTracePropagationTargets: Boolean = false,
         ): OkHttpClient {
-            whenever(hub.options).thenReturn(
-                SentryOptions().apply {
-                    dsn = "https://key@sentry.io/proj"
-                    isTraceSampling = true
-                    if (includeMockServerInTracingOrigins) {
-                        tracingOrigins.add(server.hostName)
-                    } else {
-                        tracingOrigins.add("other-api")
-                    }
+            options = SentryOptions().apply {
+                dsn = "https://key@sentry.io/proj"
+                isTraceSampling = true
+                if (includeMockServerInTracePropagationTargets) {
+                    setTracePropagationTargets(listOf(server.hostName))
+                } else if (!keepDefaultTracePropagationTargets) {
+                    setTracePropagationTargets(listOf("other-api"))
                 }
-            )
+            }
+            whenever(hub.options).thenReturn(options)
 
             sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
 
@@ -104,8 +105,28 @@ class SentryOkHttpInterceptorTest {
     }
 
     @Test
+    fun `when there is an active span and tracing origins contains default regex, adds sentry trace headers to the request`() {
+        val sut = fixture.getSut(keepDefaultTracePropagationTargets = true)
+
+        sut.newCall(getRequest()).execute()
+        val recorderRequest = fixture.server.takeRequest()
+        assertNotNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])
+        assertNotNull(recorderRequest.headers[BaggageHeader.BAGGAGE_HEADER])
+    }
+
+    @Test
     fun `when there is an active span and server is not listed in tracing origins, does not add sentry trace headers to the request`() {
-        val sut = fixture.getSut(includeMockServerInTracingOrigins = false)
+        val sut = fixture.getSut(includeMockServerInTracePropagationTargets = false)
+        sut.newCall(Request.Builder().get().url(fixture.server.url("/hello")).build()).execute()
+        val recorderRequest = fixture.server.takeRequest()
+        assertNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])
+        assertNull(recorderRequest.headers[BaggageHeader.BAGGAGE_HEADER])
+    }
+
+    @Test
+    fun `when there is an active span and server tracing origins is empty, does not add sentry trace headers to the request`() {
+        val sut = fixture.getSut()
+        fixture.options.setTracePropagationTargets(emptyList())
         sut.newCall(Request.Builder().get().url(fixture.server.url("/hello")).build()).execute()
         val recorderRequest = fixture.server.takeRequest()
         assertNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])