+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentrySpanWebClientCustomizerTest.kt
@@ -41,10 +41,11 @@ class SentrySpanWebClientCustomizerTest {
         fun getSut(isTransactionActive: Boolean, status: HttpStatus = HttpStatus.OK, throwIOException: Boolean = false, includeMockServerInTracingOrigins: Boolean = true): WebClient {
             sentryOptions = SentryOptions().apply {
                 if (includeMockServerInTracingOrigins) {
-                    tracingOrigins.add(mockServer.hostName)
+                    setTracePropagationTargets(listOf(mockServer.hostName))
                 } else {
-                    tracingOrigins.add("other-api")
+                    setTracePropagationTargets(listOf("other-api"))
                 }
+                dsn = "http://key@localhost/proj"
             }
             whenever(hub.options).thenReturn(sentryOptions)
             transaction = SentryTracer(TransactionContext("aTransaction", "op", TracesSamplingDecision(true)), hub)